@@ -7,8 +7,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"go.opentelemetry.io/otel/sdk/instrumentation"
-	semconv "go.opentelemetry.io/otel/semconv/v1.19.0"
 	"net"
 	"net/http"
 	"os"
@@ -17,6 +15,7 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel/attribute"
+	semconv "go.opentelemetry.io/otel/semconv/v1.24.0"
 	"go.opentelemetry.io/otel/trace"
 
 	"github.com/IBM/sarama"
@@ -35,6 +34,9 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
+	flagd "github.com/open-feature/go-sdk-contrib/providers/flagd/pkg"
+	"github.com/open-feature/go-sdk/openfeature"
+	otelhooks "github.com/open-feature/go-sdk-contrib/hooks/open-telemetry/pkg"
 	healthpb "google.golang.org/grpc/health/grpc_health_v1"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
@@ -111,10 +113,6 @@ func initMeterProvider() *sdkmetric.MeterProvider {
 	mp := sdkmetric.NewMeterProvider(
 		sdkmetric.WithReader(sdkmetric.NewPeriodicReader(exporter)),
 		sdkmetric.WithResource(initResource()),
-		sdkmetric.WithView(sdkmetric.NewView(
-			sdkmetric.Instrument{Scope: instrumentation.Scope{Name: "go.opentelemetry.io/contrib/google.golang.org/grpc/otelgrpc"}},
-			sdkmetric.Stream{Aggregation: sdkmetric.AggregationDrop{}},
-		)),
 	)
 	otel.SetMeterProvider(mp)
 	return mp
@@ -129,7 +127,13 @@ type checkoutService struct {
 	paymentSvcAddr        string
 	kafkaBrokerSvcAddr    string
 	pb.UnimplementedCheckoutServiceServer
-	KafkaProducerClient sarama.AsyncProducer
+	KafkaProducerClient     sarama.AsyncProducer
+	shippingSvcClient       pb.ShippingServiceClient
+	productCatalogSvcClient pb.ProductCatalogServiceClient
+	cartSvcClient           pb.CartServiceClient
+	currencySvcClient       pb.CurrencyServiceClient
+	emailSvcClient          pb.EmailServiceClient
+	paymentSvcClient        pb.PaymentServiceClient
 }
 
 func main() {
@@ -155,15 +159,42 @@ func main() {
 		log.Fatal(err)
 	}
 
+	openfeature.SetProvider(flagd.NewProvider())
+
 	tracer = tp.Tracer("checkoutservice")
 
 	svc := new(checkoutService)
+
 	mustMapEnv(&svc.shippingSvcAddr, "SHIPPING_SERVICE_ADDR")
+	c := mustCreateClient(context.Background(), svc.shippingSvcAddr)
+	svc.shippingSvcClient = pb.NewShippingServiceClient(c)
+	defer c.Close()
+
 	mustMapEnv(&svc.productCatalogSvcAddr, "PRODUCT_CATALOG_SERVICE_ADDR")
+	c = mustCreateClient(context.Background(), svc.productCatalogSvcAddr)
+	svc.productCatalogSvcClient = pb.NewProductCatalogServiceClient(c)
+	defer c.Close()
+
 	mustMapEnv(&svc.cartSvcAddr, "CART_SERVICE_ADDR")
+	c = mustCreateClient(context.Background(), svc.cartSvcAddr)
+	svc.cartSvcClient = pb.NewCartServiceClient(c)
+	defer c.Close()
+
 	mustMapEnv(&svc.currencySvcAddr, "CURRENCY_SERVICE_ADDR")
+	c = mustCreateClient(context.Background(), svc.currencySvcAddr)
+	svc.currencySvcClient = pb.NewCurrencyServiceClient(c)
+	defer c.Close()
+
 	mustMapEnv(&svc.emailSvcAddr, "EMAIL_SERVICE_ADDR")
+	c = mustCreateClient(context.Background(), svc.emailSvcAddr)
+	svc.emailSvcClient = pb.NewEmailServiceClient(c)
+	defer c.Close()
+
 	mustMapEnv(&svc.paymentSvcAddr, "PAYMENT_SERVICE_ADDR")
+	c = mustCreateClient(context.Background(), svc.paymentSvcAddr)
+	svc.paymentSvcClient = pb.NewPaymentServiceClient(c)
+	defer c.Close()
+
 	svc.kafkaBrokerSvcAddr = os.Getenv("KAFKA_SERVICE_ADDR")
 
 	if svc.kafkaBrokerSvcAddr != "" {
@@ -217,7 +248,7 @@ func (cs *checkoutService) PlaceOrder(ctx context.Context, req *pb.PlaceOrderReq
 	var err error
 	defer func() {
 		if err != nil {
-			span.AddEvent("error", trace.WithAttributes(attribute.String("exception.message", err.Error())))
+			span.AddEvent("error", trace.WithAttributes(semconv.ExceptionMessageKey.String(err.Error())))
 		}
 	}()
 
@@ -339,21 +370,20 @@ func (cs *checkoutService) prepareOrderItemsAndShippingQuoteFromCart(ctx context
 	return out, nil
 }
 
-func createClient(ctx context.Context, svcAddr string) (*grpc.ClientConn, error) {
-	return grpc.DialContext(ctx, svcAddr,
+func mustCreateClient(ctx context.Context, svcAddr string) *grpc.ClientConn {
+	c, err := grpc.DialContext(ctx, svcAddr,
 		grpc.WithTransportCredentials(insecure.NewCredentials()),
 		grpc.WithStatsHandler(otelgrpc.NewClientHandler()),
 	)
+	if err != nil {
+		log.Fatalf("could not connect to %s service, err: %+v", svcAddr, err)
+	}
+
+	return c
 }
 
 func (cs *checkoutService) quoteShipping(ctx context.Context, address *pb.Address, items []*pb.CartItem) (*pb.Money, error) {
-	conn, err := createClient(ctx, cs.shippingSvcAddr)
-	if err != nil {
-		return nil, fmt.Errorf("could not connect shipping service: %+v", err)
-	}
-	defer conn.Close()
-
-	shippingQuote, err := pb.NewShippingServiceClient(conn).
+	shippingQuote, err := cs.shippingSvcClient.
 		GetQuote(ctx, &pb.GetQuoteRequest{
 			Address: address,
 			Items:   items})
@@ -364,13 +394,7 @@ func (cs *checkoutService) quoteShipping(ctx context.Context, address *pb.Addres
 }
 
 func (cs *checkoutService) getUserCart(ctx context.Context, userID string) ([]*pb.CartItem, error) {
-	conn, err := createClient(ctx, cs.cartSvcAddr)
-	if err != nil {
-		return nil, fmt.Errorf("could not connect cart service: %+v", err)
-	}
-	defer conn.Close()
-
-	cart, err := pb.NewCartServiceClient(conn).GetCart(ctx, &pb.GetCartRequest{UserId: userID})
+	cart, err := cs.cartSvcClient.GetCart(ctx, &pb.GetCartRequest{UserId: userID})
 	if err != nil {
 		return nil, fmt.Errorf("failed to get user cart during checkout: %+v", err)
 	}
@@ -378,13 +402,7 @@ func (cs *checkoutService) getUserCart(ctx context.Context, userID string) ([]*p
 }
 
 func (cs *checkoutService) emptyUserCart(ctx context.Context, userID string) error {
-	conn, err := createClient(ctx, cs.cartSvcAddr)
-	if err != nil {
-		return fmt.Errorf("could not connect cart service: %+v", err)
-	}
-	defer conn.Close()
-
-	if _, err = pb.NewCartServiceClient(conn).EmptyCart(ctx, &pb.EmptyCartRequest{UserId: userID}); err != nil {
+	if _, err := cs.cartSvcClient.EmptyCart(ctx, &pb.EmptyCartRequest{UserId: userID}); err != nil {
 		return fmt.Errorf("failed to empty user cart during checkout: %+v", err)
 	}
 	return nil
@@ -393,15 +411,8 @@ func (cs *checkoutService) emptyUserCart(ctx context.Context, userID string) err
 func (cs *checkoutService) prepOrderItems(ctx context.Context, items []*pb.CartItem, userCurrency string) ([]*pb.OrderItem, error) {
 	out := make([]*pb.OrderItem, len(items))
 
-	conn, err := createClient(ctx, cs.productCatalogSvcAddr)
-	if err != nil {
-		return nil, fmt.Errorf("could not connect product catalog service: %+v", err)
-	}
-	defer conn.Close()
-	cl := pb.NewProductCatalogServiceClient(conn)
-
 	for i, item := range items {
-		product, err := cl.GetProduct(ctx, &pb.GetProductRequest{Id: item.GetProductId()})
+		product, err := cs.productCatalogSvcClient.GetProduct(ctx, &pb.GetProductRequest{Id: item.GetProductId()})
 		if err != nil {
 			return nil, fmt.Errorf("failed to get product #%q", item.GetProductId())
 		}
@@ -417,12 +428,7 @@ func (cs *checkoutService) prepOrderItems(ctx context.Context, items []*pb.CartI
 }
 
 func (cs *checkoutService) convertCurrency(ctx context.Context, from *pb.Money, toCurrency string) (*pb.Money, error) {
-	conn, err := createClient(ctx, cs.currencySvcAddr)
-	if err != nil {
-		return nil, fmt.Errorf("could not connect currency service: %+v", err)
-	}
-	defer conn.Close()
-	result, err := pb.NewCurrencyServiceClient(conn).Convert(ctx, &pb.CurrencyConversionRequest{
+	result, err := cs.currencySvcClient.Convert(ctx, &pb.CurrencyConversionRequest{
 		From:   from,
 		ToCode: toCurrency})
 	if err != nil {
@@ -432,13 +438,14 @@ func (cs *checkoutService) convertCurrency(ctx context.Context, from *pb.Money, 
 }
 
 func (cs *checkoutService) chargeCard(ctx context.Context, amount *pb.Money, paymentInfo *pb.CreditCardInfo) (string, error) {
-	conn, err := createClient(ctx, cs.paymentSvcAddr)
-	if err != nil {
-		return "", fmt.Errorf("failed to connect payment service: %+v", err)
-	}
-	defer conn.Close()
-
-	paymentResp, err := pb.NewPaymentServiceClient(conn).Charge(ctx, &pb.ChargeRequest{
+    paymentService := cs.paymentSvcClient
+	if cs.checkPaymentFailure(ctx) {
+        badAddress := "badAddress:50051"
+        c := mustCreateClient(context.Background(), badAddress)
+		paymentService = pb.NewPaymentServiceClient(c)
+    }
+	
+	paymentResp, err := paymentService.Charge(ctx, &pb.ChargeRequest{
 		Amount:     amount,
 		CreditCard: paymentInfo})
 	if err != nil {
@@ -470,12 +477,7 @@ func (cs *checkoutService) sendOrderConfirmation(ctx context.Context, email stri
 }
 
 func (cs *checkoutService) shipOrder(ctx context.Context, address *pb.Address, items []*pb.CartItem) (string, error) {
-	conn, err := createClient(ctx, cs.shippingSvcAddr)
-	if err != nil {
-		return "", fmt.Errorf("failed to connect email service: %+v", err)
-	}
-	defer conn.Close()
-	resp, err := pb.NewShippingServiceClient(conn).ShipOrder(ctx, &pb.ShipOrderRequest{
+	resp, err := cs.shippingSvcClient.ShipOrder(ctx, &pb.ShipOrderRequest{
 		Address: address,
 		Items:   items})
 	if err != nil {
@@ -512,9 +514,8 @@ func createProducerSpan(ctx context.Context, msg *sarama.ProducerMessage) trace.
 		trace.WithSpanKind(trace.SpanKindProducer),
 		trace.WithAttributes(
 			semconv.PeerService("kafka"),
-			semconv.NetTransportTCP,
-			semconv.MessagingSystem("kafka"),
-			semconv.MessagingDestinationKindTopic,
+			semconv.NetworkTransportTCP,
+			semconv.MessagingSystemKafka,
 			semconv.MessagingDestinationName(msg.Topic),
 			semconv.MessagingOperationPublish,
 			semconv.MessagingKafkaDestinationPartition(int(msg.Partition)),
@@ -530,4 +531,13 @@ func createProducerSpan(ctx context.Context, msg *sarama.ProducerMessage) trace.
 	}
 
 	return span
+}
+
+func (cs *checkoutService) checkPaymentFailure(ctx context.Context) bool {
+	openfeature.AddHooks(otelhooks.NewTracesHook())
+	client := openfeature.NewClient("checkout")
+	failureEnabled, _ := client.BooleanValue(
+		ctx, "paymentServiceUnreachable", false, openfeature.EvaluationContext{},
+	)
+	return failureEnabled
 }
