@@ -8,7 +8,7 @@ const { SHIPPING_SERVICE_ADDR = '' } = process.env;
 
 const client = new ShippingServiceClient(SHIPPING_SERVICE_ADDR, ChannelCredentials.createInsecure());
 
-const AdGateway = () => ({
+const ShippingGateway = () => ({
   getShippingCost(itemList: CartItem[], address: Address) {
     return new Promise<GetQuoteResponse>((resolve, reject) =>
       client.getQuote({ items: itemList, address: address }, (error, response) =>
@@ -18,4 +18,4 @@ const AdGateway = () => ({
   },
 });
 
-export default AdGateway();
+export default ShippingGateway();
