@@ -4,13 +4,22 @@ const {context, propagation, trace, metrics} = require('@opentelemetry/api');
 const cardValidator = require('simple-card-validator');
 const { v4: uuidv4 } = require('uuid');
 
+const { OpenFeature } = require('@openfeature/server-sdk');
+const { FlagdProvider} = require('@openfeature/flagd-provider');
+const flagProvider = new FlagdProvider();
+
 const logger = require('./logger');
 const tracer = trace.getTracer('paymentservice');
 const meter = metrics.getMeter('paymentservice');
 const transactionsCounter = meter.createCounter('app.payment.transactions')
 
-module.exports.charge = request => {
+module.exports.charge = async request => {
   const span = tracer.startSpan('charge');
+
+  await OpenFeature.setProviderAndWait(flagProvider);
+  if (await OpenFeature.getClient().getBooleanValue("paymentServiceFailure", false)) {
+    throw new Error("PaymentService Fail Feature Flag Enabled");
+  }
 
   const {
     creditCardNumber: number,
